% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{process_substance_info}
\alias{process_substance_info}
\title{Process results of substance_info to be more user-friendly}
\usage{
process_substance_info(sub_info)
}
\description{
Process results of substance_info to be more user-friendly
}
