% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea-query.R
\name{sea_target_vs_target}
\alias{sea_target_vs_target}
\title{compute sea score given uniprot entries for targets
for the substance fields}
\usage{
sea_target_vs_target(
  ref_target,
  query_target,
  substance_fields = c("preferred_name", "smiles", "purchasability"),
  Tc_matrix = TRUE,
  verbose
)
}
\description{
compute sea score given uniprot entries for targets
for the substance fields
}
