% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{standardize_zinc_ids}
\alias{standardize_zinc_ids}
\title{Standardize ZINC ID to have format ZINC[0-9]{12}
Specifically take all numeric values, on the right, pad with zeros to get to 12 numbers and then prepend with 'ZINC'}
\usage{
standardize_zinc_ids(zinc_ids)
}
\description{
Standardize ZINC ID to have format ZINC[0-9]{12}
Specifically take all numeric values, on the right, pad with zeros to get to 12 numbers and then prepend with 'ZINC'
}
