% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{substance_analogs}
\alias{substance_analogs}
\title{Search for analogs of the given substances}
\usage{
substance_analogs(zinc_ids, output_fields = c("zinc_id", "smiles",
  "preferred_name", "purchasability"), fingerprint = "ecfp4_fp",
  similarity = "tanimoto", threshold = 0.4, raw = FALSE, ...)
}
\description{
Search for analogs of the given substances
}
