% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{substance_info}
\alias{substance_info}
\title{Lookup info on substances by zinc_ids}
\usage{
substance_info(zinc_ids, output_fields = c("zinc_id", "preferred_name",
  "smiles", "purchasability", "features"), batch_size = length(zinc_ids),
  raw = FALSE, ...)
}
\description{
Lookup info on substances by zinc_ids
}
