% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc_tools.R
\name{tc_matrix}
\alias{tc_matrix}
\title{WARNING this may not work well}
\usage{
tc_matrix(ref_smiles, query_smiles, verbose = F,
  url_path = "tools/tc_matrix")
}
\description{
WARNING this may not work well
}

