% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{resolve_substances}
\alias{resolve_substances}
\title{Search for substances using smiles and fine grained tolerance criteria}
\usage{
resolve_substances(
  identifiers,
  allow_lookup_zincids = TRUE,
  allow_lookup_structures = TRUE,
  allow_lookup_names = TRUE,
  allow_lookup_suppliers = TRUE,
  allow_lookup_analogs = TRUE,
  match_tolerance_retired = FALSE,
  match_tolerance_charges = FALSE,
  match_tolerance_scaffolds = FALSE,
  match_tolerance_fulltext = FALSE,
  match_tolerance_multiple = FALSE,
  subsets_to_check = "all",
  raw = FALSE,
  ...
)
}
\description{
Search for substances using smiles and fine grained tolerance criteria
}
