% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc-substance.R
\name{search_for_substances}
\alias{search_for_substances}
\title{Search for substances by search terms}
\usage{
search_for_substances(search_terms, output_fields = c("zinc_id",
  "preferred_name", "smiles", "purchasability", "features"), raw, ...)
}
\description{
Search for substances by search terms
}
